requirements.txt
asgiref==3.11.0
django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.7
pyjwt==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
